feat: AccountDeleteTest is done

+++ src/main/java/ru/zalimannard/dripchip/schema/account/AccountServiceImpl.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.PositiveOrZero;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import ru.zalimannard.dripchip.exception.BadRequestException;
@@ -13,6 +14,7 @@ import ru.zalimannard.dripchip.exception.ConflictException;
 import ru.zalimannard.dripchip.exception.ForbiddenException;
 import ru.zalimannard.dripchip.exception.NotFoundException;
 import ru.zalimannard.dripchip.page.OffsetBasedPage;
+import ru.zalimannard.dripchip.security.UserSecurity;
 
 import java.util.List;
 
@@ -23,6 +25,7 @@ public class AccountServiceImpl implements AccountService {
     private final AccountRepository accountRepository;
     private final AccountMapper accountMapper;
     private final PasswordEncoder encoder;
+    private final UserSecurity userSecurity;
 
     @Override
     public AccountDto create(@Valid AccountDto accountDto) {
@@ -59,6 +62,7 @@ public class AccountServiceImpl implements AccountService {
         } catch (NotFoundException e) {
             throw new ForbiddenException();
         }
+        checkAccess(id);
 
         Account accountRequest = accountMapper.toEntity(accountDto);
         accountRequest.setId(id);
@@ -75,6 +79,7 @@ public class AccountServiceImpl implements AccountService {
         } catch (NotFoundException e) {
             throw new ForbiddenException();
         }
+        checkAccess(id);
 
         try {
             accountRepository.deleteById(id);
@@ -86,6 +91,12 @@ public class AccountServiceImpl implements AccountService {
     private void checkExist(int id) {
         if (!accountRepository.existsById(id)) {
             throw new NotFoundException("Account", String.valueOf(id));
+        }
+    }
+
+    private void checkAccess(int id) {
+        if (!userSecurity.hasUserId(SecurityContextHolder.getContext().getAuthentication(), id)) {
+            throw new ForbiddenException();
         }
     }
